// src/models/plant.ts
export interface Plant {
    isActive: boolean;
    orderDate: any;
    gardenId: string;
    plantId: string;
    plantOrigin: string;
    batchNo: string;
    cutDate: any;
    expectedRootDate: any;
    expectedHarvestDate: any;
    actualHarvestDate: any;
    productId: string;
    productName: string;
    productTrimId: string;
    productWasteId: string;
    quantity: number;
    stage: string;
    dryWeight: any;
    wetWeight: any;
    wasteWeight: any;
    trimWeight: any;
    imageURL: any;
    createdDate: any;
    createdBy: string;
    updatedDate: any;
    updatedBy: string;
}